Remove annoying i3wm mouse behavior
This is purely a personal preference of mine. If this is an undesirable change, then we can easily revert back to the old behavior we had before by making this change instead:

```
focus_follows_mouse yes
mouse_warping none
```

I was actually hoping for one of these change or the other - FWIW, I believe making things explicit and obvious to the user like this is a great practice. It is not up for me to decide what the default mouse behavior of LARBS should be moving forward, but I also think that there is absolutely no harm in explicitly defining these options in the i3wm config file.