Volume buttons support
Found my keyboard volume buttons not working in larbs (archi3). At least not always, because i feel like they worked in some scenarios on fresh install. Though at that time it was hard to tell, my audio was pretty much inexistent due to the lack of pulse & alsa. Anyway, this got it working. Sorry if it's stupid, i'm very much a newb.