Remove `exec` from `startx` in `~/.profile`
I found this useful for myself - if the `X` server fails to start,
I'd get thrown out of my session (obviously, because `exec` is used).

This meant that I had to use another user which does not use the `X` server
for troubleshooting. For me, `root` worked, but you might not have access to it,
and this could help.

I don't know if this makes the setup any less secure;
if it does - lemme know.