i3 to use slock
it could only work with '&&' and only in this order. i see that it would be logical to "slock" even if a preceding command will fail, though like i said it would only run in this order. i've tried using a script (even in this order) with " ; " it would not run. 
the audios would mute and the screen would turn off, but it would not lock.