Use `pkill` instead of `killall`
The `killall` command doesn't come shipped in all distributions (Void
Linux as an example), on the other hand `pkill` does.

Unsure what the exact intention on each call is but the `-9` signal can be
replaced with `-x` if the intention is not to end all instances.

Also note that I left `pkill -x Xorg` without the `-9` signal because
for some reason it won't function as expected and end up halting the
system (at least on Void Linux).