Added Java support for compiler script
Tested it on a project with this structure:

```
- root
 |- test
    | - Test.java
```

The user must call `compiler` from the root of the project, but that should not
be an inconvenience because usually the working directory is the project's
root. Nonetheless, it also works for java files in the current working directory,
in case someone just wants to test some code quickly.