Update sysact
this change removes the incorrect white space char next to `shutdown` and `renew` and replaces it with a regular white space char.

I'm not sure if others experience this, but I get a box character next to `shutdown` and `renew`. When I open this in my code editor and switch the font to Liberation Mono, the same box char appears. You likely don't experience this issue if you did not change your font to liberation mono.

My theory is that it's a char that doesn't render the same in Liberation Mono as it does in Luke's default font.