shortcuts: make `ls` command configurable
when merging from upstream (you) it's always a hassle
to resolve the changes in this file,
but it shouldn't be, because the only modification i make
is to change the way the `ls` command behaves.

thus by extracting the SHORTCUTS_LS_CMD command out,
even if it's edited in the file,
will no longer interfere w/ other logic of the program.

fwiw, i set mine to

```sh
ls -l --all --size --group-directories-first --human-readable --classify --color=auto
```

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>