Fix tag script mp3 issue
When you don't manually specify $total and $date with an mp3 file the eyeD3 command tries tagging with an empty variable, which causes it to error. This fixes it by only passing the parameter to the command if the variable is set. 

This pull request also adds support for the $genre and $comment variable.